Refactor quality kernels to .cxx
