+++ pyroteus/quality.py
@@ -1,8 +1,8 @@
-from pyroteus.kernel import *
 import firedrake
 from pyop2 import op2
 from firedrake.petsc import PETSc
 import ufl
+from . import kernel as kernels
 
 
 @PETSc.Log.EventDecorator("pyroteus.get_min_angles2d")
@@ -22,7 +22,7 @@ def get_min_angles2d(mesh, python=False):
     else:
         coords = mesh.coordinates
         min_angles = firedrake.Function(P0)
-        kernel = eigen_kernel(get_min_angle2d)
+        kernel = kernels.eigen_kernel(kernels.get_min_angle, 2)
         op2.par_loop(kernel, mesh.cell_set, min_angles.dat(op2.WRITE, min_angles.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return min_angles
@@ -45,7 +45,7 @@ def get_min_angles3d(mesh, python=False):
     else:
         coords = mesh.coordinates
         min_angles = firedrake.Function(P0)
-        kernel = eigen_kernel(get_min_angle3d)
+        kernel = kernels.eigen_kernel(kernels.get_min_angle, 3)
         op2.par_loop(kernel, mesh.cell_set, min_angles.dat(op2.WRITE, min_angles.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return min_angles
@@ -68,7 +68,7 @@ def get_areas2d(mesh, python=False):
     else:
         coords = mesh.coordinates
         areas = firedrake.Function(P0)
-        kernel = eigen_kernel(get_area2d)
+        kernel = kernels.eigen_kernel(kernels.get_area, 2)
         op2.par_loop(kernel, mesh.cell_set, areas.dat(op2.WRITE, areas.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return areas
@@ -91,7 +91,7 @@ def get_volumes3d(mesh, python=False):
     else:
         coords = mesh.coordinates
         volumes = firedrake.Function(P0)
-        kernel = eigen_kernel(get_volume3d)
+        kernel = kernels.eigen_kernel(kernels.get_volume, 3)
         op2.par_loop(kernel, mesh.cell_set, volumes.dat(op2.WRITE, volumes.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return volumes
@@ -186,8 +186,9 @@ def get_aspect_ratios2d(mesh, python=False):
     else:
         coords = mesh.coordinates
         aspect_ratios = firedrake.Function(P0)
-        kernel = eigen_kernel(get_aspect_ratio2d)
-        op2.par_loop(kernel, mesh.cell_set, aspect_ratios.dat(op2.WRITE, aspect_ratios.cell_node_map()),
+        kernel = kernels.eigen_kernel(kernels.get_aspect_ratio, 2)
+        op2.par_loop(kernel, mesh.cell_set, 
+                     aspect_ratios.dat(op2.WRITE, aspect_ratios.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return aspect_ratios
 
@@ -209,8 +210,9 @@ def get_aspect_ratios3d(mesh, python=False):
     else:
         coords = mesh.coordinates
         aspect_ratios = firedrake.Function(P0)
-        kernel = eigen_kernel(get_aspect_ratio3d)
-        op2.par_loop(kernel, mesh.cell_set, aspect_ratios.dat(op2.WRITE, aspect_ratios.cell_node_map()),
+        kernel = kernels.eigen_kernel(kernels.get_aspect_ratio, 3)
+        op2.par_loop(kernel, mesh.cell_set, 
+                     aspect_ratios.dat(op2.WRITE, aspect_ratios.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return aspect_ratios
 
@@ -232,7 +234,7 @@ def get_eskews2d(mesh, python=False):
     else:
         coords = mesh.coordinates
         eskews = firedrake.Function(P0)
-        kernel = eigen_kernel(get_eskew2d)
+        kernel = kernels.eigen_kernel(kernels.get_eskew, 2)
         op2.par_loop(kernel, mesh.cell_set, eskews.dat(op2.WRITE, eskews.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return eskews
@@ -255,7 +257,7 @@ def get_eskews3d(mesh, python=False):
     else:
         coords = mesh.coordinates
         eskews = firedrake.Function(P0)
-        kernel = eigen_kernel(get_eskew3d)
+        kernel = kernels.eigen_kernel(kernels.get_eskew, 3)
         op2.par_loop(kernel, mesh.cell_set, eskews.dat(op2.WRITE, eskews.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return eskews
@@ -278,7 +280,7 @@ def get_skewnesses2d(mesh, python=False):
     else:
         coords = mesh.coordinates
         skews = firedrake.Function(P0)
-        kernel = eigen_kernel(get_skewness2d)
+        kernel = kernels.eigen_kernel(kernels.get_skewness, 2)
         op2.par_loop(kernel, mesh.cell_set, skews.dat(op2.WRITE, skews.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return skews
@@ -327,8 +329,9 @@ def get_scaled_jacobians2d(mesh, python=False):
     else:
         coords = mesh.coordinates
         scaled_jacobians = firedrake.Function(P0)
-        kernel = eigen_kernel(get_scaled_jacobian2d)
-        op2.par_loop(kernel, mesh.cell_set, scaled_jacobians.dat(op2.WRITE, scaled_jacobians.cell_node_map()),
+        kernel = kernels.eigen_kernel(kernels.get_scaled_jacobian, 2)
+        op2.par_loop(kernel, mesh.cell_set, 
+                     scaled_jacobians.dat(op2.WRITE, scaled_jacobians.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return scaled_jacobians
 
@@ -350,8 +353,9 @@ def get_scaled_jacobians3d(mesh, python=False):
     else:
         coords = mesh.coordinates
         scaled_jacobians = firedrake.Function(P0)
-        kernel = eigen_kernel(get_scaled_jacobian3d)
-        op2.par_loop(kernel, mesh.cell_set, scaled_jacobians.dat(op2.WRITE, scaled_jacobians.cell_node_map()),
+        kernel = kernels.eigen_kernel(kernels.get_scaled_jacobian, 3)
+        op2.par_loop(kernel, mesh.cell_set, 
+                     scaled_jacobians.dat(op2.WRITE, scaled_jacobians.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
     return scaled_jacobians
 
@@ -375,7 +379,7 @@ def get_quality_metrics2d(mesh, metric, python=False):
     else:
         coords = mesh.coordinates
         quality = firedrake.Function(P0)
-        kernel = eigen_kernel(get_metric2d)
+        kernel = kernels.eigen_kernel(kernels.get_metric, 2)
         op2.par_loop(kernel, mesh.cell_set, quality.dat(op2.WRITE, quality.cell_node_map()),
                      metric.dat(op2.READ, metric.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
@@ -401,7 +405,7 @@ def get_quality_metrics3d(mesh, metric, python=False):
     else:
         coords = mesh.coordinates
         quality = firedrake.Function(P0)
-        kernel = eigen_kernel(get_metric3d)
+        kernel = kernels.eigen_kernel(kernels.get_metric, 3)
         op2.par_loop(kernel, mesh.cell_set, quality.dat(op2.WRITE, quality.cell_node_map()),
                      metric.dat(op2.READ, metric.cell_node_map()),
                      coords.dat(op2.READ, coords.cell_node_map()))
